fixed 2801a5e from https://gitee.com/wangxiaojia1204/arkcompiler_ets_runtime/pulls/4452 Fix data type PART 1
Issue:https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I7LZJR

+++ ecmascript/ts_types/ts_type_table.cpp
@@ -18,7 +18,7 @@
 namespace panda::ecmascript {
 JSHandle<JSTaggedValue> TSTypeTable::GetExportValueTable(JSThread *thread, JSHandle<TSTypeTable> typeTable)
 {
-    int index = static_cast<int>(typeTable->GetLength()) - 1;
+    uint32_t index = typeTable->GetLength() - 1;
     JSHandle<JSTaggedValue> exportValueTable(thread, typeTable->Get(index));
     return exportValueTable;
 }
